+++ CHANGES.rst
@@ -6,7 +6,8 @@ CHANGES
 
 * Fix handling of *shared* ID3 metadata across audio files when *only* global options are provided.
   For example, only giving ``--artist <ARTIST>`` without any other audio configuration file to match audio files
-  against caused many ``AttributeError`` and incorrect application of specified tags to *all* files.
+  against caused many ``AttributeError`` and incorrect application of specified tags to *all* files
+  (fixes `#1 <https://github.com/fmigneault/aiu/issues/1>`_).
 
 `1.0.0 <https://github.com/fmigneault/aiu/tree/1.0.0>`_ (2021-03-02)
 ------------------------------------------------------------------------------------